refactoring components
Added components folder

Seperated components within own files and added them to compenents
