+++ .vs/VSWorkspaceState.json
@@ -2,8 +2,9 @@
   "ExpandedNodes": [
     "",
     "\\public",
-    "\\src"
+    "\\src",
+    "\\src\\components"
   ],
-  "SelectedNode": "\\package.json",
+  "SelectedNode": "\\src\\components\\Box.js",
   "PreviewInSolutionExplorer": false
 }
